+++ src/lib/Scenes/Artwork/__tests__/Artwork-tests.tsx
@@ -23,4 +23,18 @@ describe("Artwork", () => {
     const component = mount(<Artwork artwork={ArtworkFixture} relay={{ environment: {} } as RelayRefetchProp} />)
     expect(component.find(ArtworkHeader).length).toEqual(1)
   })
+
+  it("refetches on re-appear", () => {
+    const refetchMock = jest.fn()
+    const component = mount(
+      <Artwork
+        artwork={ArtworkFixture as any}
+        relay={({ environment: {}, refetch: refetchMock } as unknown) as RelayRefetchProp}
+        isVisible
+      />
+    )
+    component.setProps({ isVisible: false })
+    component.setProps({ isVisible: true })
+    expect(refetchMock).toHaveBeenCalled()
+  })
 })